# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zero-Click CRM API
  version: 1.0.0
  description: >-
    Documentación de la API para el backend Zero-Click CRM. Incluye ingestión multicanal,
    procesamiento de audio/video, consultas CRM, chat asistido por IA, alertas, notas,
    knowledge base y endpoints administrativos.
  contact:
    name: Equipo Relay
servers:
  - url: http://localhost:4000
    description: Servidor de desarrollo local
  - url: https://your-deployment-url.example
    description: Servidor de producción (actualiza esta URL al desplegar)
...
paths:
...
  /api/ingest/email:
    post:
      summary: Ingesta de email (mock o real)
      description: Normaliza un email a una interacción CRM. Si no se especifica body o incluye `generate=true`, genera datos mock automáticamente.
      tags: [Ingest]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEmailRequest'
      responses:
        '200':
          description: Email ingerido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/ingest/slack:
    post:
      summary: Ingesta de mensaje de Slack (mock o real)
      tags: [Ingest]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestSlackRequest'
      responses:
        '200':
          description: Mensaje ingerido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/ingest/whatsapp:
    post:
      summary: Ingesta de mensaje de WhatsApp (mock o real)
      tags: [Ingest]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestWhatsAppRequest'
      responses:
        '200':
          description: Mensaje ingerido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/ingest/audio:
    post:
      summary: Sube audio a Supabase Storage y dispara procesamiento automático
      tags: [Ingest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestAudioVideoRequest'
      responses:
        '200':
          description: Audio cargado y en proceso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/ingest/video:
    post:
      summary: Sube video a Supabase Storage y dispara procesamiento automático
      tags: [Ingest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestAudioVideoRequest'
      responses:
        '200':
          description: Video cargado y en proceso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/jobs/audio:
    post:
      summary: Procesa audio inline con Gemini y genera interacción
      tags: [Jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobAudioRequest'
      responses:
        '200':
          description: Audio procesado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/jobs/video:
    post:
      summary: Procesa video inline con Gemini (audio_only o multimodal)
      tags: [Jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobVideoRequest'
      responses:
        '200':
          description: Video procesado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  schemas:
...
    IngestEmailRequest:
      type: object
      properties:
        from:
          type: string
          example: 'Patricia Peña <patricia.pena@techsmart.com>'
        to:
          type: string
          example: ventas@miempresa.com
        subject:
          type: string
        body:
          type: string
        date:
          type: string
          format: date-time
        company:
          type: string
        attachments:
          type: array
          items:
            type: object
        metadata:
          type: object
        generate:
          type: boolean
          description: Si es true ignora el resto del body y genera datos mock
    IngestSlackRequest:
      type: object
      properties:
        user:
          type: object
        channel:
          type: object
        text:
          type: string
          example: 'Hola equipo, tenemos una nueva oportunidad con TechSmart...'
        ts:
          type: string
        thread_ts:
          type: string
          nullable: true
        metadata:
          type: object
        generate:
          type: boolean
    IngestWhatsAppRequest:
      type: object
      properties:
        from:
          type: string
          example: '+521234567890'
        to:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        contactName:
          type: string
        email:
          type: string
        company:
          type: string
        media:
          type: object
          nullable: true
        metadata:
          type: object
        generate:
          type: boolean
    IngestAudioVideoRequest:
      type: object
      properties:
        bucket:
          type: string
        filePath:
          type: string
        localPath:
          type: string
        base64:
          type: string
        mimeType:
          type: string
        process:
          type: boolean
          default: true
    StorageUploadResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
        bucket:
          type: string
        path:
          type: string
        url:
          type: string
        willProcess:
          type: boolean
    IngestResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        interactionId:
          type: string
        contactId:
          type: string
          nullable: true
        companyId:
          type: string
          nullable: true
        generated:
          type: boolean
        data:
          type: object
          nullable: true
    JobAudioRequest:
      type: object
      required: [audio]
      properties:
        audio:
          type: object
          required: [base64, mimeType]
          properties:
            base64:
              type: string
            mimeType:
              type: string
        source:
          type: string
          default: call
        metadata:
          type: object
          default: {}
    JobVideoRequest:
      allOf:
        - $ref: '#/components/schemas/JobAudioRequest'
        - type: object
          properties:
            video:
              type: object
              required: [base64, mimeType]
              properties:
                base64:
                  type: string
                mimeType:
                  type: string
            analysis:
              type: string
              enum: [audio_only, video]
              default: audio_only
    JobResponse:
      type: object
      properties:
        ok:
          type: boolean
        jobId:
          type: string
        status:
          type: string
        data:
          type: object
        contact:
          type: object
          nullable: true
...
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
...
    ServiceUnavailable:
      description: Servicio externo no configurado o no disponible (Gemini/Embeddings)
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string